Minor UI changes and dummy proofing

+++ main.py
@@ -9,10 +9,14 @@ import glob
 
 my_library = None
 
+
 def main():
     print('Howdy!')
-    print(my_library.get_library_name())
-    
+    if my_library.get_library_name() == 'temp':
+        pass
+    else:
+        print(my_library.get_library_name())
+
     # Loops until user correctly enters for new or existing library
     while my_library.get_library_name() == 'temp':
         choose_new_existing()
@@ -20,6 +24,7 @@ def main():
     while True:
         display_actions(my_library.get_library_name())
         user_action = input("Choice: ").strip().lower()
+        print()
         match user_action:
             # Prints all books
             case '1':
@@ -47,13 +52,16 @@ def main():
                         else:
                             my_library.print_all_books()
                             index = int(input(
-                                "If you would like to exit delete book mode, enter any alphabetical letter.\nWhich book would you like to remove? (Type the ID): ").strip())
+                                "To exit enter [x]\nWhich book would you like to remove? (Enter the ID): ").strip())
+                            
                             if get_yn_input(f'Are you sure you would like to delete {my_library.seek_book(index)}'):
                                 my_library.remove_book(index)
                             else:
-                                print("Exiting delete book mode...")
+                                print("Canceling deletion...")
+                                print()
                                 break
                     except ValueError:
+                        print("Exiting delete mode...")
                         print()
                         break
 
@@ -71,7 +79,7 @@ def main():
                 my_library.save_books(f'{my_library.get_library_name()}')
             case 'b':
                 confirmation = get_yn_input(
-                    "This will delete all newly added changes. Please ensure you have saved the data to file. Continue? (y/n) ")
+                    "This will delete all newly added changes. Please ensure you have saved the data to file. Would you still like to go back? (y/n)")
                 # If yes
                 if confirmation:
                     my_library.wipe()
@@ -83,9 +91,12 @@ def main():
                 elif not confirmation:
                     pass
             case 'x':
-                print("Exiting program...")
-                delete_temp_csv()
-                break
+                confirmation = get_yn_input(
+                    "This will delete all newly added changes. Please ensure you have saved the data to file. Would you still like to go back? (y/n)")
+                if confirmation:
+                    print("Exiting program...")
+                    delete_temp_csv()
+                    break
             case _:
                 print("Invalid choice. Please try again")
 
@@ -110,6 +121,7 @@ def display_actions(library_name):
 def get_yn_input(prompt):
     while True:
         user_input = input(prompt).strip().lower()
+        print()
         if user_input == "y" or user_input == "yes":
             return True
         elif user_input == "n" or user_input == "no":
@@ -152,7 +164,7 @@ def create_list_libraries():
 def choose_new_existing():
     while True:
         create_new = input(
-            "Please enter [n] for creating a new library or [e] for viewing an existing library.\nOr press [x] to quit. ").strip().lower()
+            "Enter [n] to create new library or [e] for viewing an existing library.\nPress [x] to quit. ").strip().lower()
 
         # Create a list of all pre-existing libraries
         existing_libraries = create_list_libraries()
@@ -162,9 +174,9 @@ def choose_new_existing():
 
             library_name = input(
                 "\nWhat would you like to name your library?\nOr press [x] to quit. ").strip()
-            
+
             duplicate_flag = False
-            
+
             if library_name == "x" or library_name == "X":
                 # Go back to the start of the first while loop
                 break
@@ -172,7 +184,7 @@ def choose_new_existing():
                 for library in existing_libraries:
                     if library == library_name:
                         duplicate_flag = True
-                    
+
                 if duplicate_flag:
                     print("Invalid name. This library already exists.")
                 else:
@@ -196,6 +208,7 @@ def choose_new_existing():
                 # Ask the user, which library they want to view
                 get_existing_library = input(
                     "\nWhat is the name of the library you wish to view?\n").strip().lower()
+                print()
 
                 if f'{get_existing_library}' in existing_libraries:
                     my_library.set_library_name(get_existing_library)
@@ -212,16 +225,20 @@ def choose_new_existing():
             print('Invalid input, please try again')
 
 # Delete the temp file
+
+
 def delete_temp_csv():
     try:
         os.remove('temp.csv')
     except:
         pass
 
+
 def initialize_library():
     global my_library
     my_library = Library('temp')
-    
+
+
 if __name__ == "__main__":
     initialize_library()
     main()
